uint16 is not enough to store 17 flags, uint32 will be enough

+++ src/svutil.h
@@ -195,7 +195,7 @@ struct FuseGene{
     std::string tgene;   ///< gene name of 3' part of fusion gene
     std::string tend;    ///< tgene 3' or 5' fused
     std::string tstrand; ///< hgene + or - strand fused
-    uint16_t status;      ///< mask to show fusion status 1:gene,2:normal,4:hot,8:common,16:indb,32:mirror
+    uint32_t status;      ///< mask to show fusion status 1:gene,2:normal,4:hot,8:common,16:indb,32:mirror
 
     /** FuseGene constructor */
     FuseGene(){
